+++ eelbrain/pipeline.py
@@ -5,7 +5,7 @@ from ._experiment.preprocessing import RawSource, RawFilter, RawICA, RawMaxwell,
 from ._experiment.epochs import ContinuousEpoch, PrimaryEpoch, SecondaryEpoch, SuperEpoch
 from ._experiment.groups import Group, SubGroup
 from ._experiment.parc import SubParc, CombinationParc, FreeSurferParc, FSAverageParc, SeededParc, IndividualSeededParc
-from ._experiment.test_def import ANOVA, TTestOneSample, TTestIndependent, TTestRelated, TContrastRelated, TwoStageTest
+from ._experiment.test_def import ANOVA, TTestOneSample, TTestIndependent, TTestRelated, TContrastRelated, TwoStageTest, ROITestResult, ROI2StageResult
 from ._experiment.variable_def import EvalVar, GroupVar, LabelVar
 
 # backwards compatibility